+++ phantom-dsl/src/test/scala/com/outworkers/phantom/tables/TestDatabase.scala
@@ -107,12 +107,8 @@ class TestDatabase(
   // table to test a schema bug for using RootSelectBlockOps on select projections with no other clauses
   object userSchema extends UserSchemaTable with Connector
   object verizonSchema extends VerizonSchema with Connector
-  object schemaBugSecondaryIndex extends SchemaBugSecondaryIndex with Connector
 }
 
-class SecondaryIndexOnlyDatabase(
-  override val connector: CassandraConnection
-) extends Database[SecondaryIndexOnlyDatabase](connector)
 
 object Connector {
   val default: CassandraConnection = connectors.ContactPoint.local
@@ -126,19 +122,6 @@ object Connector {
         replication eqs SimpleStrategy.replication_factor(1)
       )
     )
-
-  val specialTests = connectors.ContactPoint.local
-    .withClusterBuilder(_.withSocketOptions(
-      new SocketOptions()
-        .setConnectTimeoutMillis(20000)
-        .setReadTimeoutMillis(20000)
-    )
-    ).noHeartbeat().keySpace(
-    KeySpace("phantom_special").ifNotExists().`with`(
-      replication eqs SimpleStrategy.replication_factor(1)
-    )
-  )
-
 }
 
 object TestDatabase extends TestDatabase(Connector.default)
