+++ phantom-dsl/src/test/scala/com/outworkers/phantom/tables/dbs/SecondaryIndexOnlyDatabase.scala
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2013 - 2017 Outworkers Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.outworkers.phantom.tables.dbs
+
+import com.datastax.driver.core.SocketOptions
+import com.outworkers.phantom.connectors
+import com.outworkers.phantom.connectors.CassandraConnection
+import com.outworkers.phantom.database.Database
+import com.outworkers.phantom.dsl._
+import com.outworkers.phantom.tables.bugs.SchemaBugSecondaryIndex
+
+object TestConnector {
+
+  val specialTests = connectors.ContactPoint.local
+    .withClusterBuilder(_.withSocketOptions(
+      new SocketOptions()
+        .setConnectTimeoutMillis(20000)
+        .setReadTimeoutMillis(20000)
+    )
+    ).noHeartbeat().keySpace(
+    KeySpace("phantom_special").ifNotExists().`with`(
+      replication eqs SimpleStrategy.replication_factor(1)
+    )
+  )
+}
+
+class SecondaryIndexOnlyDatabase(
+  override val connector: CassandraConnection
+) extends Database[SecondaryIndexOnlyDatabase](connector) {
+  object schemaBugSecondaryIndex extends SchemaBugSecondaryIndex with Connector
+}
+
+object SecondaryIndexOnlyDatabase extends SecondaryIndexOnlyDatabase(TestConnector.specialTests)
